move code closer together
