+++ deployables/extension/src/panel/pages/$activeAccountId/transactions/ContractAddress.tsx
@@ -1,10 +1,9 @@
+import type { ContractInfo } from '@/state'
 import { EXPLORER_URL } from '@zodiac/chains'
 import type { HexAddress } from '@zodiac/schema'
 import { AddressInput, GhostLinkButton } from '@zodiac/ui'
-import { getAddress } from 'ethers'
 import { SquareArrowOutUpRight } from 'lucide-react'
 import type { ChainId } from 'ser-kit'
-import type { ContractInfo } from '../../../../utils/abi'
 
 interface Props {
   chainId: ChainId
@@ -13,22 +12,18 @@ interface Props {
 }
 
 export const ContractAddress = ({ chainId, address, contractInfo }: Props) => {
-  const explorerUrl = EXPLORER_URL[chainId]
-
-  const checksumAddress = address && getAddress(address)
-
   return (
     <AddressInput
       readOnly
       label="Contract"
       description={contractInfo?.name}
-      value={checksumAddress as HexAddress}
+      value={address}
       action={
         <GhostLinkButton
           openInNewWindow
           iconOnly
           size="small"
-          to={`${explorerUrl}/search?q=${address}`}
+          to={`${EXPLORER_URL[chainId]}/search?q=${address}`}
           icon={SquareArrowOutUpRight}
         >
           Show in block explorer
