+++ deployables/extension/src/panel/providers-ui/MockBrowserProvider.ts
@@ -1,0 +1,12 @@
+import type { BrowserProvider } from 'ethers'
+import { vi, type MockedFunction } from 'vitest'
+
+export class MockBrowserProvider {
+  getTransactionReceipt: MockedFunction<
+    BrowserProvider['getTransactionReceipt']
+  >
+
+  constructor() {
+    this.getTransactionReceipt = vi.fn().mockResolvedValue({})
+  }
+}
