+++ deployables/extension/src/panel/providers-ui/useDecodeTransactions.ts
@@ -1,7 +1,8 @@
 import { useAccount } from '@/accounts'
+import { sentry } from '@/sentry'
 import { decodeTransaction, useDispatch, usePendingTransactions } from '@/state'
 import { useEffect } from 'react'
-import { fetchContractInfo } from '../utils'
+import { fetchContractInfo } from './fetchContractInfo'
 
 export const useDecodeTransactions = () => {
   const pendingTransactions = usePendingTransactions()
@@ -14,9 +15,13 @@ export const useDecodeTransactions = () => {
         continue
       }
 
-      fetchContractInfo(transaction.to, chainId).then((contractInfo) => {
-        dispatch(decodeTransaction({ id: transaction.id, contractInfo }))
-      })
+      fetchContractInfo(transaction.to, chainId)
+        .then((contractInfo) => {
+          dispatch(decodeTransaction({ id: transaction.id, contractInfo }))
+        })
+        .catch((error) => {
+          sentry.captureException(error)
+        })
     }
   }, [chainId, dispatch, pendingTransactions])
 }
